+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @agentmark/ui-components
 
+## 0.2.0
+
+### Minor Changes
+
+- 82ca057: Add status message in trace span for failure
+
 ## 0.1.0
 
 ### Minor Changes